requirements.txt
aniso8601==9.0.1
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
coverage==7.1.0
cryptography==37.0.2
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==4.11.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
MarkupSafe==2.1.2
newrelic==6.4.4.161
packaging==23.0
pluggy==1.0.0
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.19.3
pytest==7.2.1
python-dotenv==0.20.0
pytz==2022.7.1
requests==2.27.1
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.12.1
